2 (White background)
Every line of code is the same as "1" except for the color of the background. In this case, I put it as "0", so there is no black value whatsoever. As you can see, the opacity of the square is the same as well, and because of the nature of the opacity effect, the square appears as such.